Create groups.txt
group names in Pi-hole. 

As-much a like-for-like in Domain tagging / category. So lists can be grouped and set to clients in Pi-hole in a content filtering environment.

some lists maybe on multiple groups

this is my personal preference for group names. Each will have their own preference on how they setup groups in Pi-hole.